Add exec_tail to the test framework
